requirements.txt
streamlit==1.39.0
numpy==1.26.4
pillow==10.4.0
torch==2.3.0+cpu
torchvision==0.18.0+cpu 